fix(emit): the real, cross-cutting loop-variable I32 boxing bug
Founder: "can we fix the loop-variable I32 boxing bug?" -- the gap
STDLIB.md's own linalg/base4 sections had documented as "genuinely
cross-cutting... not attempted" since 2026-08-21.

Root cause: NODE_NUMBER always reported "double" for every numeric
literal (no int/float distinction), so a loop/let binding seeded from
a whole-number literal like `0` was C-typed double. That's silently
fine for arithmetic, but vec_push_/vec_set_at_'s own scalar-boxing
decision picks vec_box_f64 (an 8-byte double cell) the instant such a
value is pushed into an I32-typed Vec, while a later typed read does
*(int *) on that same cell -- real memory-level corruption, not just a
wrong value.

Fix, in two parts:
1. NODE_NUMBER reports "int" for a literal with no '.'/'e'/'E' in its
   own text, "double" otherwise. Doesn't change any already-correct C
   arithmetic (a bare literal's own C text already governed real
   division/promotion, independent of this internal type string) --
   only downstream decisions (declared variable types, box function
   choice) that were wrong before.
2. A `let` binding, assigned exactly once, is safe with this change
   immediately. A `loop` binding can be reassigned via `recur` on every
   iteration, so it additionally needs `loop_body_int_safe()`: after
   every binding is known, walk the loop's own body (mirroring
   emit_loop_tail's real if/cond/when/do/let/match traversal, stopping
   at a nested loop) and downgrade an "int"-typed binding back to
   "double" -- today's pre-existing, safe default -- unless EVERY real
   recur update for it also resolves to "int" via the real emit_expr
   type inference (not a separate, necessarily-incomplete hand-rolled
   subset -- a first draft using a narrow +/-/* recognizer regressed
   set.prn's own fnv1a-hash, which updates via bit-xor/*; reusing real
   emit_expr fixed it by construction).

Verified: base4/vector.prn's `dot` and base4/matrix.prn's `matmul` --
both previously had their own tests asserting the CONFIRMED-buggy
behavior as a regression gate -- now produce the real, hand-traced
correct values via the correct int* cast; both tests and both files'
own doc comments updated. linalg.prn's own array/set!-based matmul/
transpose remain separately unverified (array.prn's own [i j] index-
Vec-literal path may have additional, unrelated bugs -- a real,
separate, deeper investigation, not claimed fixed here).

make test: 347/347 (2 pre-existing tests updated: a literal now boxes
via vec_box_i32 not vec_box_f64; a loop's own result type now
correctly resolves to int not double -- both are the bug's own fix,
not new breakage). Every other real, buildable test target re-run
clean: base4/awk/ami/ber/bstree/datetime/dns/dtmf/g711/git/http-*/
io-mmap/json/linalg-sparse/mag-gematria/mixforge/multifile/net-proxy/
net-rawsocket/pentest-*/process/rtp/set/sip-*/textmate-loader/v16-*/
webdriver/wire/yaml/selfhost-*/construct-split -- zero regressions.

backlog: ✓ loop-variable-i32-boxing-bug


session: sess-20260905-0720-ec33e7c5

+++ tests/test_base4_gen.c
@@ -56,14 +56,14 @@ int base4_subtract(int a __attribute__((unused)), int b __attribute__((unused)))
 int base4_iterate(int start __attribute__((unused)), int (*op)(int, int) __attribute__((unused)), int steps __attribute__((unused))) {
     int __loop_result_0 __attribute__((unused));
     int state = start;
-    double i = 0;
+    int i = 0;
     while (1) {
         if ((i >= steps)) {
         __loop_result_0 = state;
         break;
         } else {
         int __recur_tmp_0 = op(state, start);
-        double __recur_tmp_1 = (i + 1);
+        int __recur_tmp_1 = (i + 1);
         state = __recur_tmp_0;
         i = __recur_tmp_1;
         continue;
@@ -73,9 +73,9 @@ int base4_iterate(int start __attribute__((unused)), int (*op)(int, int) __attri
 }
 
 int base4_cycle_length(int start __attribute__((unused)), int (*op)(int, int) __attribute__((unused))) {
-    double __loop_result_1 __attribute__((unused));
+    int __loop_result_1 __attribute__((unused));
     int state = op(start, start);
-    double i = 1;
+    int i = 1;
     while (1) {
         if ((state == start)) {
         __loop_result_1 = i;
@@ -86,7 +86,7 @@ int base4_cycle_length(int start __attribute__((unused)), int (*op)(int, int) __
         break;
         } else {
         int __recur_tmp_0 = op(state, start);
-        double __recur_tmp_1 = (i + 1);
+        int __recur_tmp_1 = (i + 1);
         state = __recur_tmp_0;
         i = __recur_tmp_1;
         continue;
