+++ tests/test_mag_gematria_gen.c
@@ -112,14 +112,14 @@ int is_valid_i32_text_(char * s __attribute__((unused))) {
     return 0;
     } else {
     int __loop_result_0 __attribute__((unused));
-    double i = (starts_with_sign_(s) ? 1 : 0);
+    int i = (starts_with_sign_(s) ? 1 : 0);
     int ok = 1;
     while (1) {
         if ((i >= n)) {
         __loop_result_0 = (ok && (n > (starts_with_sign_(s) ? 1 : 0)));
         break;
         } else {
-        double __recur_tmp_0 = (i + 1);
+        int __recur_tmp_0 = (i + 1);
         int __recur_tmp_1 = (ok && is_digit_(char_at(s, i)));
         i = __recur_tmp_0;
         ok = __recur_tmp_1;
@@ -139,8 +139,8 @@ char * concat(char * a __attribute__((unused)), char * b __attribute__((unused))
 Vec split(char * s __attribute__((unused)), char * sep __attribute__((unused)), Arena *dest __attribute__((unused))) {
     Vec result __attribute__((unused)) = vec_new(dest);
     Vec __loop_result_1 __attribute__((unused));
-    double start = 0;
-    double i = 0;
+    int start = 0;
+    int i = 0;
     int n = length(s);
     while (1) {
         if ((i >= n)) {
@@ -150,16 +150,16 @@ Vec split(char * s __attribute__((unused)), char * sep __attribute__((unused)), 
         } else {
         if ((char_at(s, i) == char_at(sep, 0))) {
         (void)(vec_push_(&(result), substring(s, start, i, dest)));
-        double __recur_tmp_0 = (i + 1);
-        double __recur_tmp_1 = (i + 1);
+        int __recur_tmp_0 = (i + 1);
+        int __recur_tmp_1 = (i + 1);
         int __recur_tmp_2 = n;
         start = __recur_tmp_0;
         i = __recur_tmp_1;
         n = __recur_tmp_2;
         continue;
         } else {
-        double __recur_tmp_0 = start;
-        double __recur_tmp_1 = (i + 1);
+        int __recur_tmp_0 = start;
+        int __recur_tmp_1 = (i + 1);
         int __recur_tmp_2 = n;
         start = __recur_tmp_0;
         i = __recur_tmp_1;
@@ -186,16 +186,16 @@ int za_digit(int code __attribute__((unused)), int grp_len __attribute__((unused
 }
 
 int fingerprint_az(char * s __attribute__((unused)), int grp_len __attribute__((unused))) {
-    double __loop_result_2 __attribute__((unused));
-    double i = 0;
-    double acc = 0;
+    int __loop_result_2 __attribute__((unused));
+    int i = 0;
+    int acc = 0;
     while (1) {
         if ((i >= length(s))) {
         __loop_result_2 = acc;
         break;
         } else {
-        double __recur_tmp_0 = (i + 1);
-        double __recur_tmp_1 = ((acc * 10) + az_digit(char_at(s, i), grp_len));
+        int __recur_tmp_0 = (i + 1);
+        int __recur_tmp_1 = ((acc * 10) + az_digit(char_at(s, i), grp_len));
         i = __recur_tmp_0;
         acc = __recur_tmp_1;
         continue;
@@ -205,16 +205,16 @@ int fingerprint_az(char * s __attribute__((unused)), int grp_len __attribute__((
 }
 
 int fingerprint_za(char * s __attribute__((unused)), int grp_len __attribute__((unused))) {
-    double __loop_result_3 __attribute__((unused));
-    double i = 0;
-    double acc = 0;
+    int __loop_result_3 __attribute__((unused));
+    int i = 0;
+    int acc = 0;
     while (1) {
         if ((i >= length(s))) {
         __loop_result_3 = acc;
         break;
         } else {
-        double __recur_tmp_0 = (i + 1);
-        double __recur_tmp_1 = ((acc * 10) + za_digit(char_at(s, i), grp_len));
+        int __recur_tmp_0 = (i + 1);
+        int __recur_tmp_1 = ((acc * 10) + za_digit(char_at(s, i), grp_len));
         i = __recur_tmp_0;
         acc = __recur_tmp_1;
         continue;
